Fixed issue #18936: User count in group is not OK after deleting a user
