+++ application/controllers/UserManagementController.php
@@ -1,7 +1,6 @@
 <?php
 
 use LimeSurvey\Models\Services\UserManager;
-use LimeSurvey\Models\Services\UserManagementHelper;
 
 //LSYii_Controller
 /**
@@ -410,8 +409,8 @@ class UserManagementController extends LSBaseController
             }
         }
 
-        $userManagementHelper = new UserManagementHelper();
-        $result = $userManagementHelper->deleteUser($userId);
+        $userManager = new UserManager();
+        $result = $userManager->deleteUser($userId);
         $messages = array_merge($messages, $result->getRawMessages());
 
         return App()->getController()->renderPartial('/admin/super/_renderJson', [
