+++ application/models/services/UserManager.php
@@ -5,6 +5,8 @@ namespace LimeSurvey\Models\Services;
 use LSWebUser;
 use Permission;
 use User;
+use LimeSurvey\Datavalueobjects\OperationResult;
+use LimeSurvey\Datavalueobjects\TypedMessage;
 
 /**
  * Service class for managing users and their permissions
@@ -18,11 +20,11 @@ class UserManager
     private $targetUser;
 
     /**
-     * @param LSWebUser $managingUser
+     * @param LSWebUser|null $managingUser
      * @param User|null $targetUser
      */
     public function __construct(
-        LSWebUser $managingUser,
+        LSWebUser $managingUser = null,
         User $targetUser = null
     ) {
         $this->managingUser = $managingUser;
@@ -82,5 +84,58 @@ class UserManager
                 Permission::model()->hasGlobalPermission('users', 'update', $this->managingUser->id)
                 && $this->targetUser->parent_id == $this->managingUser->id
             );
+    }
+
+    /**
+     * Deletes the user with the given id.
+     * @param int $userId
+     * @return OperationResult
+     */
+    public function deleteUser($userId)
+    {
+        $messages = [];
+
+        $siteAdminName = \User::model()->findByPk(1)->users_name;
+
+        $transaction = \Yii::app()->db->beginTransaction();
+        try {
+            // Transfer any User Groups owned by this user to site's admin
+            $userGroupsTranferred = \UserGroup::model()->updateAll(['owner_id' => 1], 'owner_id = :owner_id', [':owner_id' => $userId]);
+            if ($userGroupsTranferred) {
+                $messages[] = new TypedMessage(sprintf(gT("All of the user's user groups were transferred to %s."), $siteAdminName), 'success');
+            }
+
+            // Transfer any Participants owned by this user to site's admin
+            $participantsTranferred = \Participant::model()->updateAll(['owner_uid' => 1], 'owner_uid = :owner_uid', [':owner_uid' => $userId]);
+            if ($participantsTranferred) {
+                $messages[] = new TypedMessage(sprintf(gT("All participants owned by this user were transferred to %s."), $siteAdminName), 'success');
+            }
+
+            // Remove from groups
+            $userAndGroupRelations = \UserInGroup::model()->findAllByAttributes(['uid' => $userId]);
+            if (count($userAndGroupRelations)) {
+                foreach ($userAndGroupRelations as $userAndGroupRelation) {
+                    $userAndGroupRelation->delete();
+                }
+            }
+
+            // TODO: User permissions should be deleted also...
+
+            // Delete the user
+            $user = \User::model()->findByPk($userId);
+            $success = $user->delete();
+            if (!$success) {
+                $messages = [new TypedMessage(gT("User could not be deleted."), 'error')];
+                $transaction->rollback();
+            } else {
+                $messages[] = new TypedMessage(gT("User successfully deleted."), 'success');
+                $transaction->commit();
+            }
+        } catch (\Exception $e) {
+            $transaction->rollback();
+            $messages = [new TypedMessage(gT("An error occurred while deleting the user."), 'error')];
+            $success = false;
+        }
+        return new OperationResult($success, $messages);
     }
 }
